Move QueryContext to the parent module.
